Update files and methods
For example, we can now in stable category compute
lift, colift, compute a morphism that lift and colift at the same time
which will help us later implement Isomorphism(To,From)CanonicalExactTriangle.

since we have now lifts and colifts, we can decide if a morphism is
split epi or split mono and hence if it is isomorphism.

We are able now to also compute not only one lift/colift but all of them.